requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.3.0
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pymysql==1.1.0
python-dotenv==1.0.0
jinja2==3.1.2
psycopg2-binary==2.9.9 